refactor(frr): extract render validation belt; finish the split
Move the render-side value sanitization/validation belt out of
policy_render.go into a cohesive sibling render_validate.go (#6424):
sanitizeFRRValue plus the validRouterID / validClusterID / validBGPOrigin
predicates that keep a leniently-loaded or peer-synced malformed value
out of the generated frr.conf. This is the final extraction of the
god-renderer split.

The functions are moved whole and unchanged; only the now-unused
"strconv" (residual) and "strings" (validation belt) imports are dropped
to satisfy the compiler. The residual policy_render.go header is rewritten
to describe its remaining responsibility (policy-statement -> route-map
rendering, community classification, redist-alias guard) and to index the
sibling files. No logic or render output change; the emitted frr.conf is
byte-identical.

Validation: go build ./pkg/frr, gofmt, go vet clean. policy_render.go
drops from 969 to 887 LOC (from 2,470 at the start of the split).
